requirements.txt
streamlit
openai
langchain
langchain-community
python-dotenv
